Initial commit
This code is based on Train Log by Zomis and utilizes the FLib framework for GUIs and events.  The original code has been extensively modified to work with Space Exploration cargo rockets.

NOTE: This mod is not compatible with the official release of Space Exploration.  A major API interface update is required before they will work together.